feat: refactor paymentTest
