+++ tests/Feature/OrderTest.php
@@ -8,11 +8,8 @@ use App\Actions\OrderProduct\StoreUpdateOrderProductActions;
 use App\Constants\Constants;
 use App\Models\Product;
 use App\Models\User;
-use App\PaymentGateways\Placetopay;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
-use Illuminate\Support\Facades\Config;
 
 class OrderTest extends TestCase
 {
@@ -67,20 +64,4 @@ class OrderTest extends TestCase
         $response->assertStatus(200);
     }
 
-    public function test_payment_geteway()
-    {
-        $arrayPay = [
-            'reference' => 1,
-            'total' => 1000,
-            'returnUrl' =>  Constants::URL_RETURN_PLACETOPAY . '/1',
-            'description' => Constants::DESCRIPTION_PLACETOPAY . " 1",
-            'currency' => Constants::CURRENCY
-        ];
-
-        $paymentGeteway = new Placetopay();
-
-        $response = $paymentGeteway->createSession($arrayPay);
-
-        $this->assertSame((string)$response['status']['status'], "OK");
-    }
 }
